+++ CMakeLists.txt
@@ -68,8 +68,8 @@ set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/Observable)
 set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/Sampler)
 set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/Optimizer)
 set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/Stats)
-set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/UnitaryTimeEvolution)
-set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/UnitaryTimeEvolution/TimeStepper)
+set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/UnitaryDynamics)
+set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/UnitaryDynamics/TimeStepper)
 set (NETKET_SOURCES ${NETKET_SOURCES} ${NETKET_SOURCE_DIR}/Utils)
 
 add_executable(netket ${NETKET_SOURCES})
