+++ NetKet/UnitaryDynamics/time_evolution.hpp
@@ -6,8 +6,9 @@
 #include "Graph/graph.hpp"
 #include "Hamiltonian/hamiltonian.hpp"
 #include "Hamiltonian/MatrixWrapper/matrix_wrapper.hpp"
-#include "UnitaryTimeEvolution/TimeStepper/time_stepper.hpp"
 #include "Utils/all_utils.hpp"
+
+#include "TimeStepper/time_stepper.hpp"
 
 namespace netket
 {
